requirements.txt
requests
coinbase
python-dotenv
psycopg2
fpdf